Move Subject into Facet.Core.Type.
