+++ src/Facet/Elab.hs
@@ -14,9 +14,6 @@ module Facet.Elab
   -- * Errors
 , pushSpan
 , Err(..)
-, Subject(..)
-, unSubject
-, subjectType
 , ErrReason(..)
 , UnifyErrReason(..)
 , err
@@ -187,18 +184,6 @@ data Err = Err
   , subst     :: Subst Type
   , callStack :: CallStack
   }
-
-data Subject
-  = SK Kind
-  | ST Type
-
-unSubject :: (Kind -> a) -> (Type -> a) -> Subject -> a
-unSubject fk ft = \case
-  SK k -> fk k
-  ST t -> ft t
-
-subjectType :: Subject -> Maybe Type
-subjectType = unSubject (const empty) pure
 
 data ErrReason
   = FreeVariable QName
